+++ lib/runners/processor/checkstyle.rb
@@ -3,21 +3,17 @@ module Runners
     include Java
 
     Schema = StrongJSON.new do
-      let :runner_config, Schema::BaseConfig.base.update_fields { |fields|
+      let :runner_config, Schema::BaseConfig.java.update_fields { |fields|
         fields.merge!({
-                        config: string?,
-                        dir: optional(enum(string, array(string))),
-                        exclude: optional(enum(
-                                            string,
-                                            array(enum(
-                                                    string,
-                                                    object(pattern: string),
-                                                    object(string: string)
-                                                  ))
-                                          )),
-                        ignore: array?(string),
-                        properties: string?
-                      })
+          config: string?,
+          dir: enum?(string, array(string)),
+          exclude: enum?(
+            string,
+            array(enum(string, object(pattern: string), object(string: string))),
+          ),
+          ignore: array?(string),
+          properties: string?,
+        })
       }
 
       let :issue, object(
@@ -26,6 +22,16 @@ module Runners
     end
 
     register_config_schema(name: :checkstyle, schema: Schema.runner_config)
+
+    def setup
+      begin
+        install_jvm_deps
+      rescue UserError => exn
+        return Results::Failure.new(guid: guid, message: exn.message)
+      end
+
+      yield
+    end
 
     def analyze(changes)
       delete_unchanged_files(changes, only: ["*.java"])