+++ test/smokes/checkstyle/deps_not_default/Foo.java
@@ -1,0 +1,5 @@
+class Foo {
+  public Function<String, Integer> test() {
+    return (string) -> 1;
+  }
+}