+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Updated tools:
 Misc:
 
 - Download JVM dependencies via Gradle [#1352](https://github.com/sider/runners/pull/1352)
+- Install 3rd-party JVM dependencies via `sider.yml` [#1361](https://github.com/sider/runners/pull/1361)
 
 ## 0.31.0